# Trucking-Game-API/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Trucking Software Public API",
		"description": "API for trucking software web application",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:3001/api/v1",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/auth/updateDetails": {
			"patch": {
				"description": "Update user details for a single user",
				"summary": "Update user details for a single user",
				"tags": ["auth"],
				"operationId": "updateUserDetails",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Update user details success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": "true"
										},
										"updatedUser": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Update user details fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/users": {
			"get": {
				"description": "Get all users",
				"summary": "Get all users",
				"operationId": "getAllUsers",
				"tags": ["users"],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer",
							"example": 10
						},
						"required": false,
						"description": "The limit of documents to return. Defaults to 10."
					},
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "integer",
							"example": 1
						},
						"required": false,
						"description": "The page of documents to return, will determine the amount of documents to skip when querying. Defaults to 1."
					}
				],
				"responses": {
					"200": {
						"description": "Get all users success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"users": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/User"
											}
										},
										"totalUsers": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get all users fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"parameters": [
				{
					"in": "path",
					"name": "userId",
					"schema": {
						"type": "string"
					},
					"required": true,
					"description": "Id of a specific user"
				}
			],
			"get": {
				"description": "Get a single user by id",
				"summary": "Get a single user by id",
				"operationId": "getUserById",
				"tags": ["users"],
				"responses": {
					"200": {
						"description": "Get user by id success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"user": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get user by id fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"patch": {
				"description": "Update a single user by id",
				"summary": "Update a single user by id",
				"operationId": "updateUser",
				"tags": ["users"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string",
										"example": "updated name"
									},
									"email": {
										"type": "string",
										"example": "updated@email.com"
									},
									"phone": {
										"type": "string",
										"example": "847-038-3429"
									}
								},
								"required": ["name", "email", "phone"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Update user success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"updatedUser": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Update user fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/users/admin": {
			"get": {
				"description": "Get all admin users",
				"summary": "Get all admin users",
				"operationId": "getAllAdmin",
				"tags": ["users"],
				"responses": {
					"200": {
						"description": "Get all admin success",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"users": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/User"
											}
										},
										"totalUsers": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Get all admin fail",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized error response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": "false"
					},
					"error": {
						"type": "string"
					}
				}
			},
...
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "integer"
					},
					"phone": {
						"type": "string"
					},
					"profileImage": {
						"type": "string"
					},
					"hasPaid": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	]
}